Сделал обработку ошибок в проекте

+++ opinions_app/api_views.py
@@ -3,18 +3,28 @@ from flask import jsonify, request
 from . import app, db
 from .models import Opinion
 from .views import random_opinion
+from .error_handlers import InvalidAPIUsage
 
 
 @app.route('/api/opinions/<int:id>/', methods=['GET'])
 def get_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     return jsonify({'opinion': opinion.to_dict()}), 200
 
 
 @app.route('/api/opinions/<int:id>/', methods=['PATCH'])
 def update_opinion(id):
     data = request.get_json()
-    opinion = Opinion.query.get_or_404(id)
+    if (
+        'text' in data and
+        Opinion.query.filter_by(text=data['text']).first() is not None
+    ):
+        raise InvalidAPIUsage('Такое мнение уже есть в БД.')
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     opinion.title = data.get('title', opinion.title)
     opinion.text = data.get('text', opinion.text)
     opinion.source = data.get('source', opinion.source)
@@ -25,7 +35,9 @@ def update_opinion(id):
 
 @app.route('/api/opinions/<int:id>/', methods=['DELETE'])
 def delete_opinion(id):
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if opinion is None:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
     return '', 204
@@ -41,6 +53,10 @@ def get_opinions():
 @app.route('/api/opinions/', methods=['POST'])
 def add_opinion():
     data = request.get_json()
+    if 'title' not in data or 'text' not in data:
+        raise InvalidAPIUsage('В запросе отсутствуют обязательные поля!')
+    if Opinion.query.filter_by(text=data['text']).first() is not None:
+        raise InvalidAPIUsage('Такое мнение уже есть в БД.')
     opinion = Opinion()
     opinion.from_dict(data)
     db.session.add(opinion)
@@ -51,4 +67,6 @@ def add_opinion():
 @app.route('/api/get-random-opinion/', methods=['GET'])
 def get_random_opinion():
     opinion = random_opinion()
-    return jsonify({'opinion': opinion.to_dict()}), 200
+    if opinion is not None:
+        return jsonify({'opinion': opinion.to_dict()}), 200
+    raise InvalidAPIUsage('В БД нет мнений', 404)
